Split stormwater.rs (518 lines) into stormwater/ subdirectory
Split crates/simulation/src/stormwater.rs into focused modules
under stormwater/ for better maintainability:
- calculations.rs: constants and pure functions (73 lines)
- types.rs: StormwaterGrid resource and StormwaterPlugin (64 lines)
- systems.rs: update_stormwater system (134 lines)
- tests.rs: all unit tests (265 lines)
- mod.rs: thin facade with re-exports (15 lines)
