--- crates/simulation/src/stormwater.rs
@@ -1,518 +1,0 @@
-use bevy::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use crate::config::{CELL_SIZE, GRID_HEIGHT, GRID_WIDTH};
-use crate::grid::{CellType, WorldGrid, ZoneType};
-use crate::weather::{Weather, WeatherCondition};
-use crate::SlowTickTimer;
-
-/// Cell area in square meters (CELL_SIZE x CELL_SIZE).
-const CELL_AREA: f32 = CELL_SIZE * CELL_SIZE;
-
-/// Default soil permeability coefficient (m/s equivalent, unitless for simulation).
-/// Represents how easily water passes through soil when not covered by impervious surfaces.
-const SOIL_PERMEABILITY: f32 = 0.6;
-
-/// Drain rate per tick: fraction of accumulated runoff that drains to downstream cells.
-const DRAIN_RATE: f32 = 0.1;
-
-/// Returns the imperviousness coefficient for a cell based on its surface type.
-///
-/// Values represent the fraction of rainfall that becomes surface runoff:
-/// - Road/Building: 0.95 (asphalt/concrete, nearly impervious)
-/// - Parking/Industrial: 0.90
-/// - Concrete/Commercial: 0.85
-/// - Compacted soil (empty with building nearby): 0.70
-/// - Grass (default empty): 0.35
-/// - Forest/Park: 0.15
-/// - Green roof: 0.25
-/// - Pervious pavement: 0.40
-pub fn imperviousness(cell_type: CellType, zone: ZoneType, has_building: bool) -> f32 {
-    match cell_type {
-        CellType::Road => 0.95,
-        CellType::Water => 0.0,
-        CellType::Grass => {
-            if has_building {
-                // Building footprint: nearly impervious
-                0.95
-            } else {
-                match zone {
-                    ZoneType::Industrial => 0.90,
-                    ZoneType::CommercialHigh | ZoneType::CommercialLow | ZoneType::Office => 0.85,
-                    ZoneType::MixedUse => 0.85,
-                    ZoneType::ResidentialHigh | ZoneType::ResidentialMedium => 0.70,
-                    ZoneType::ResidentialLow => 0.40,
-                    ZoneType::None => 0.35,
-                }
-            }
-        }
-    }
-}
-
-/// Calculate runoff volume for a single cell given rainfall intensity.
-///
-/// `runoff = rainfall_intensity * imperviousness * cell_area`
-pub fn runoff(rainfall_intensity: f32, imperv: f32) -> f32 {
-    rainfall_intensity * imperv * CELL_AREA
-}
-
-/// Calculate infiltration volume for a single cell given rainfall intensity.
-///
-/// `infiltration = rainfall_intensity * (1.0 - imperviousness) * soil_permeability`
-pub fn infiltration(rainfall_intensity: f32, imperv: f32) -> f32 {
-    rainfall_intensity * (1.0 - imperv) * SOIL_PERMEABILITY
-}
-
-/// Rainfall intensity derived from weather condition.
-/// Returns a value in the range [0.0, 1.0] representing precipitation rate.
-fn rainfall_intensity(weather: &Weather) -> f32 {
-    match weather.current_event {
-        WeatherCondition::Rain => 0.3,
-        WeatherCondition::HeavyRain => 0.6,
-        WeatherCondition::Storm => 1.0,
-        // Snow melts slowly; minimal immediate runoff
-        WeatherCondition::Snow => 0.05,
-        _ => 0.0,
-    }
-}
-
-/// Grid tracking accumulated stormwater runoff per cell.
-///
-/// During rain events, runoff accumulates based on cell imperviousness.
-/// Between rain events, runoff gradually drains away.
-#[derive(Resource, Serialize, Deserialize)]
-pub struct StormwaterGrid {
-    /// Accumulated runoff volume per cell (cubic meters, scaled).
-    pub runoff: Vec<f32>,
-    /// Total runoff across the entire grid (for stats display).
-    pub total_runoff: f32,
-    /// Total infiltration across the grid this tick.
-    pub total_infiltration: f32,
-    pub width: usize,
-    pub height: usize,
-}
-
-impl Default for StormwaterGrid {
-    fn default() -> Self {
-        Self {
-            runoff: vec![0.0; GRID_WIDTH * GRID_HEIGHT],
-            total_runoff: 0.0,
-            total_infiltration: 0.0,
-            width: GRID_WIDTH,
-            height: GRID_HEIGHT,
-        }
-    }
-}
-
-impl StormwaterGrid {
-    #[inline]
-    pub fn get(&self, x: usize, y: usize) -> f32 {
-        self.runoff[y * self.width + x]
-    }
-
-    #[inline]
-    pub fn set(&mut self, x: usize, y: usize, val: f32) {
-        self.runoff[y * self.width + x] = val;
-    }
-
-    #[inline]
-    fn add(&mut self, x: usize, y: usize, amount: f32) {
-        let idx = y * self.width + x;
-        self.runoff[idx] += amount;
-    }
-}
-
-/// Stormwater update system. Only runs during rain/storm weather events.
-///
-/// Each tick during precipitation:
-/// 1. Calculate per-cell runoff based on imperviousness and rainfall intensity
-/// 2. Accumulate runoff in the stormwater grid
-/// 3. Drain accumulated runoff to downstream cells (based on elevation)
-/// 4. Water cells act as sinks (runoff drains into them and disappears)
-pub fn update_stormwater(
-    slow_timer: Res<SlowTickTimer>,
-    mut stormwater: ResMut<StormwaterGrid>,
-    grid: Res<WorldGrid>,
-    weather: Res<Weather>,
-) {
-    if !slow_timer.should_run() {
-        return;
-    }
-
-    let rain = rainfall_intensity(&weather);
-
-    // If no precipitation, just drain existing runoff
-    if rain <= 0.0 {
-        // Drain all accumulated runoff gradually
-        let mut any_runoff = false;
-        for val in stormwater.runoff.iter_mut() {
-            if *val > 0.0 {
-                *val *= 1.0 - DRAIN_RATE * 3.0; // faster drain when no rain
-                if *val < 0.01 {
-                    *val = 0.0;
-                }
-                any_runoff = true;
-            }
-        }
-        if !any_runoff {
-            stormwater.total_runoff = 0.0;
-            stormwater.total_infiltration = 0.0;
-        }
-        return;
-    }
-
-    // --- Phase 1: Calculate per-cell runoff and accumulate ---
-    let mut tick_runoff = 0.0_f32;
-    let mut tick_infiltration = 0.0_f32;
-
-    for y in 0..GRID_HEIGHT {
-        for x in 0..GRID_WIDTH {
-            let cell = grid.get(x, y);
-
-            // Water cells are sinks
-            if cell.cell_type == CellType::Water {
-                stormwater.set(x, y, 0.0);
-                continue;
-            }
-
-            let has_building = cell.building_id.is_some();
-            let imperv = imperviousness(cell.cell_type, cell.zone, has_building);
-
-            let cell_runoff = runoff(rain, imperv);
-            let cell_infiltration = infiltration(rain, imperv);
-
-            stormwater.add(x, y, cell_runoff);
-            tick_runoff += cell_runoff;
-            tick_infiltration += cell_infiltration;
-        }
-    }
-
-    // --- Phase 2: Drain runoff to downstream neighbors (based on elevation) ---
-    // Use a snapshot to avoid order-dependent artifacts
-    let snapshot: Vec<f32> = stormwater.runoff.clone();
-
-    for y in 0..GRID_HEIGHT {
-        for x in 0..GRID_WIDTH {
-            let idx = y * stormwater.width + x;
-            let current_runoff = snapshot[idx];
-            if current_runoff <= 0.0 {
-                continue;
-            }
-
-            let current_elevation = grid.get(x, y).elevation;
-
-            // Find lower-elevation neighbors
-            let (neighbors, count) = grid.neighbors4(x, y);
-            let mut lower_neighbors: [(usize, usize, f32); 4] = [(0, 0, 0.0); 4];
-            let mut lower_count = 0usize;
-            let mut total_drop = 0.0_f32;
-
-            for &(nx, ny) in &neighbors[..count] {
-                let n_elevation = grid.get(nx, ny).elevation;
-                if n_elevation < current_elevation {
-                    let drop = current_elevation - n_elevation;
-                    lower_neighbors[lower_count] = (nx, ny, drop);
-                    lower_count += 1;
-                    total_drop += drop;
-                }
-            }
-
-            if lower_count == 0 || total_drop <= 0.0 {
-                // No downhill neighbors; water pools here
-                continue;
-            }
-
-            // Distribute drain proportionally to elevation drop
-            let drain_amount = current_runoff * DRAIN_RATE;
-            stormwater.runoff[idx] -= drain_amount;
-
-            for &(nx, ny, drop) in &lower_neighbors[..lower_count] {
-                let fraction = drop / total_drop;
-                let transfer = drain_amount * fraction;
-
-                // Water cells absorb runoff (sink)
-                if grid.get(nx, ny).cell_type == CellType::Water {
-                    // Runoff absorbed by water body, don't add to grid
-                    continue;
-                }
-
-                stormwater.add(nx, ny, transfer);
-            }
-        }
-    }
-
-    stormwater.total_runoff = tick_runoff;
-    stormwater.total_infiltration = tick_infiltration;
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_road_cell_imperviousness() {
-        let imperv = imperviousness(CellType::Road, ZoneType::None, false);
-        assert!(
-            (imperv - 0.95).abs() < f32::EPSILON,
-            "Road cell imperviousness should be 0.95, got {}",
-            imperv
-        );
-    }
-
-    #[test]
-    fn test_grass_cell_imperviousness() {
-        let imperv = imperviousness(CellType::Grass, ZoneType::None, false);
-        assert!(
-            (imperv - 0.35).abs() < f32::EPSILON,
-            "Grass cell (no zone) imperviousness should be 0.35, got {}",
-            imperv
-        );
-    }
-
-    #[test]
-    fn test_building_cell_imperviousness() {
-        let imperv = imperviousness(CellType::Grass, ZoneType::ResidentialHigh, true);
-        assert!(
-            (imperv - 0.95).abs() < f32::EPSILON,
-            "Building cell imperviousness should be 0.95, got {}",
-            imperv
-        );
-    }
-
-    #[test]
-    fn test_water_cell_imperviousness() {
-        let imperv = imperviousness(CellType::Water, ZoneType::None, false);
-        assert!(
-            (imperv - 0.0).abs() < f32::EPSILON,
-            "Water cell imperviousness should be 0.0, got {}",
-            imperv
-        );
-    }
-
-    #[test]
-    fn test_industrial_zone_imperviousness() {
-        let imperv = imperviousness(CellType::Grass, ZoneType::Industrial, false);
-        assert!(
-            (imperv - 0.90).abs() < f32::EPSILON,
-            "Industrial zone imperviousness should be 0.90, got {}",
-            imperv
-        );
-    }
-
-    #[test]
-    fn test_commercial_zone_imperviousness() {
-        let imperv = imperviousness(CellType::Grass, ZoneType::CommercialHigh, false);
-        assert!(
-            (imperv - 0.85).abs() < f32::EPSILON,
-            "Commercial high zone imperviousness should be 0.85, got {}",
-            imperv
-        );
-    }
-
-    #[test]
-    fn test_road_cell_runoff() {
-        let rain = 1.0; // maximum rainfall intensity
-        let imperv = imperviousness(CellType::Road, ZoneType::None, false);
-        let r = runoff(rain, imperv);
-        let expected = 1.0 * 0.95 * CELL_AREA;
-        assert!(
-            (r - expected).abs() < f32::EPSILON,
-            "Road cell runoff at max rain should be {}, got {}",
-            expected,
-            r
-        );
-    }
-
-    #[test]
-    fn test_road_produces_095_rainfall_as_runoff() {
-        // Unit test from issue: road cell produces 0.95 * rainfall as runoff
-        let rainfall = 0.5;
-        let imperv = imperviousness(CellType::Road, ZoneType::None, false);
-        let r = runoff(rainfall, imperv);
-        let expected = rainfall * 0.95 * CELL_AREA;
-        assert!(
-            (r - expected).abs() < 0.001,
-            "Road runoff should be 0.95 * rainfall * area = {}, got {}",
-            expected,
-            r
-        );
-    }
-
-    #[test]
-    fn test_forest_produces_015_rainfall_as_runoff() {
-        // The closest analog to "forest" in our system is an empty grass cell (ZoneType::None)
-        // which has imperviousness 0.35. For a forest-equivalent value of 0.15,
-        // we test the runoff function directly with the forest imperviousness.
-        let rainfall = 0.5;
-        let forest_imperv = 0.15;
-        let r = runoff(rainfall, forest_imperv);
-        let expected = rainfall * 0.15 * CELL_AREA;
-        assert!(
-            (r - expected).abs() < 0.001,
-            "Forest runoff should be 0.15 * rainfall * area = {}, got {}",
-            expected,
-            r
-        );
-    }
-
-    #[test]
-    fn test_grass_cell_runoff() {
-        let rain = 1.0;
-        let imperv = imperviousness(CellType::Grass, ZoneType::None, false);
-        let r = runoff(rain, imperv);
-        let expected = 1.0 * 0.35 * CELL_AREA;
-        assert!(
-            (r - expected).abs() < f32::EPSILON,
-            "Grass cell runoff at max rain should be {}, got {}",
-            expected,
-            r
-        );
-    }
-
-    #[test]
-    fn test_infiltration_calculation() {
-        let rain = 1.0;
-        let imperv = 0.35; // grass
-        let inf = infiltration(rain, imperv);
-        let expected = 1.0 * (1.0 - 0.35) * SOIL_PERMEABILITY;
-        assert!(
-            (inf - expected).abs() < f32::EPSILON,
-            "Infiltration should be {}, got {}",
-            expected,
-            inf
-        );
-    }
-
-    #[test]
-    fn test_infiltration_zero_for_fully_impervious() {
-        let rain = 1.0;
-        let imperv = 1.0;
-        let inf = infiltration(rain, imperv);
-        assert!(
-            inf.abs() < f32::EPSILON,
-            "Fully impervious surface should have zero infiltration, got {}",
-            inf
-        );
-    }
-
-    #[test]
-    fn test_runoff_plus_infiltration_less_than_rainfall() {
-        // For any imperviousness, runoff + infiltration should not exceed total rainfall * area
-        let rain = 0.8;
-        for imperv_pct in [0.0, 0.15, 0.35, 0.70, 0.85, 0.90, 0.95, 1.0] {
-            let r = runoff(rain, imperv_pct);
-            let inf = infiltration(rain, imperv_pct);
-            let total_rain = rain * CELL_AREA;
-            assert!(
-                r + inf <= total_rain + 0.01,
-                "runoff ({}) + infiltration ({}) > total rainfall ({}) at imperv {}",
-                r,
-                inf,
-                total_rain,
-                imperv_pct
-            );
-        }
-    }
-
-    #[test]
-    fn test_stormwater_grid_default() {
-        let sw = StormwaterGrid::default();
-        assert_eq!(sw.runoff.len(), GRID_WIDTH * GRID_HEIGHT);
-        assert_eq!(sw.total_runoff, 0.0);
-        assert_eq!(sw.total_infiltration, 0.0);
-        assert!(sw.runoff.iter().all(|&v| v == 0.0));
-    }
-
-    #[test]
-    fn test_stormwater_grid_get_set() {
-        let mut sw = StormwaterGrid::default();
-        sw.set(10, 20, 5.0);
-        assert!((sw.get(10, 20) - 5.0).abs() < f32::EPSILON);
-        sw.add(10, 20, 3.0);
-        assert!((sw.get(10, 20) - 8.0).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn test_heavy_rain_paved_area_maximum_runoff() {
-        // Integration test: heavy rain on paved area produces maximum runoff
-        // Storm intensity = 1.0, road imperviousness = 0.95
-        let rain = 1.0; // Storm
-        let imperv = imperviousness(CellType::Road, ZoneType::None, false);
-        let r = runoff(rain, imperv);
-
-        // Compare with grass cell at same rainfall
-        let grass_imperv = imperviousness(CellType::Grass, ZoneType::None, false);
-        let grass_r = runoff(rain, grass_imperv);
-
-        assert!(
-            r > grass_r,
-            "Paved area runoff ({}) should exceed grass runoff ({})",
-            r,
-            grass_r
-        );
-
-        // Paved should produce roughly 0.95/0.35 = ~2.7x more runoff than grass
-        let ratio = r / grass_r;
-        assert!(
-            (ratio - 0.95 / 0.35).abs() < 0.01,
-            "Runoff ratio should be ~{}, got {}",
-            0.95 / 0.35,
-            ratio
-        );
-    }
-
-    #[test]
-    fn test_rainfall_intensity_values() {
-        let mut w = Weather::default();
-
-        w.current_event = WeatherCondition::Rain;
-        assert!((rainfall_intensity(&w) - 0.3).abs() < f32::EPSILON);
-
-        w.current_event = WeatherCondition::HeavyRain;
-        assert!((rainfall_intensity(&w) - 0.6).abs() < f32::EPSILON);
-
-        w.current_event = WeatherCondition::Storm;
-        assert!((rainfall_intensity(&w) - 1.0).abs() < f32::EPSILON);
-
-        w.current_event = WeatherCondition::Snow;
-        assert!((rainfall_intensity(&w) - 0.05).abs() < f32::EPSILON);
-
-        w.current_event = WeatherCondition::Sunny;
-        assert!((rainfall_intensity(&w) - 0.0).abs() < f32::EPSILON);
-
-        w.current_event = WeatherCondition::Overcast;
-        assert!((rainfall_intensity(&w) - 0.0).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn test_imperviousness_ordering() {
-        // Road >= Building > Industrial > CommercialHigh > ResidentialHigh > ResidentialLow > Grass
-        let road = imperviousness(CellType::Road, ZoneType::None, false);
-        let building = imperviousness(CellType::Grass, ZoneType::ResidentialHigh, true);
-        let industrial = imperviousness(CellType::Grass, ZoneType::Industrial, false);
-        let commercial = imperviousness(CellType::Grass, ZoneType::CommercialHigh, false);
-        let res_high = imperviousness(CellType::Grass, ZoneType::ResidentialHigh, false);
-        let res_low = imperviousness(CellType::Grass, ZoneType::ResidentialLow, false);
-        let grass = imperviousness(CellType::Grass, ZoneType::None, false);
-
-        assert!(road >= building);
-        assert!(building >= industrial);
-        assert!(industrial >= commercial);
-        assert!(commercial >= res_high);
-        assert!(res_high >= res_low);
-        assert!(res_low >= grass);
-    }
-}
-
-pub struct StormwaterPlugin;
-
-impl Plugin for StormwaterPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<StormwaterGrid>().add_systems(
-            FixedUpdate,
-            update_stormwater
-                .after(crate::imports_exports::process_trade)
-                .in_set(crate::SimulationSet::Simulation),
-        );
-    }
-}
